Split todo view from project view
